test: keep a snapshotting suite in its own file
CI caught what local runs could not: on Node 20 every test passed and the
job still failed, with 51 obsolete snapshots. A snapshot belonging to a
block skipped inside a file that still runs has no test to match, so jest
reports it obsolete, and the rows from Node 20 up pass `--ci` without the
`-u` that quietly prunes it on the older ones. A whole-file skip never had
the problem because the pack was not read at all.

So the CSS, swc-html and image suites go back to their own files and their
own file-level skips, and `jest.config.js` says why. What stays merged is
what takes no snapshot: the watch and filesystem-cache cases on `generate`,
`imageminGenerate` beside the generator it belongs to, `terserOptions` on
the option it is an alias of, the worker on `parallel`, and the query and
version cases on `minify`. `imageminNormalizeConfig` joins the image file,
which shares its gate.

Sixteen files rather than thirteen, and the block-level gate is kept only
where it is sound. Verified by forcing each gate off and re-running: 553
then 389 tests, no obsolete snapshot either time.
